symmetrics/requirements.txt
cytoolz==0.12.3
eciespy==0.4.1
eth-hash==0.7.0
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
falcon==3.1.3
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
Naked==0.1.32
optigatrust==1.3.8
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyserial==3.5
PyYAML==6.0.1
requests==2.32.3
shellescape==3.8.1
tinyec==0.4.0
toolz==0.12.1
typing_extensions==4.12.0
urllib3==2.2.1